docs: updated the readme
